% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustfd_endpoints.R
\name{endpoint_exists}
\alias{endpoint_exists}
\title{Tests if an endpoint is known}
\usage{
endpoint_exists(endpoint)
}
\arguments{
\item{endpoint}{character vector}
}
\value{
logical matching input size
}
\description{
See \code{\link[=ustfd_tables]{ustfd_tables()}} for known endpoints.
}
\examples{
library(ustfd)
endpoint_exists('v2/accounting/od/debt_to_penny')

}
\seealso{
Other ustfd_user: 
\code{\link{ustfd_all_pages}()},
\code{\link{ustfd_datasets}()},
\code{\link{ustfd_query}()},
\code{\link{ustfd_simple}()},
\code{\link{ustfd_table_columns}()},
\code{\link{ustfd_tables}()}
}
\concept{ustfd_user}
