% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustfd_url.R
\name{ustfd_url}
\alias{ustfd_url}
\title{Generate URL To Acces US Treasury Fiscal Data API}
\usage{
ustfd_url(query)
}
\arguments{
\item{query}{required list}
}
\value{
a httr url object
}
\description{
\code{ustfd_url()} will generate a URL suitable for querying the Fiscal Data API.
}
\examples{

library(ustfd)
ustfd_url(ustfd_query('/v1/accounting/dts/dts_table_4'))


}
\seealso{
Other ustfd_low_level: 
\code{\link{ustfd_json_response}()},
\code{\link{ustfd_request}()},
\code{\link{ustfd_response_meta_object}()},
\code{\link{ustfd_response_payload}()}
}
\concept{ustfd_low_level}
